move : 채팅 서비스 Entity 파일명 변경
