+++ apps/chat/src/chat/chat.module.ts
@@ -4,8 +4,8 @@ import { MongooseModule } from '@nestjs/mongoose';
 import { ChatController } from './chat.controller';
 import { ChatService } from './chat.service';
 import { ChatGateway } from './chat.gateway';
-import { CHAT_LIST_SCHEMA, ChatList } from './entity/chat-list.entity';
-import { CHAT_ROOM_SCHEMA, ChatRoom } from './entity/chat-room.entity';
+import { CHAT_LIST_SCHEMA, ChatList } from './entities/chat-list.entity';
+import { CHAT_ROOM_SCHEMA, ChatRoom } from './entities/chat-room.entity';
 
 @Module({
   imports: [
