+++ apps/chat/src/chat/entities/chat-list.entity.ts
