+++ apps/chat/src/chat/entities/chat-room.entity.ts
