+++ apps/chat/src/chat/entities/chat.entity.ts
